Trailing Zeroes in Binary Representation Of Number
Trailing Zeroes in Binary Representation of a Given no. or input no. :-
Most of the students know how to write a program to find a trailing zero , but when there is a twist come in a question or problem statement like find the trailing zeros in Binary Representation of a input number then everyone losses the hope . but now I come up with a super optimized solution for such problems with having O(1) Time Complexity (Constant time complexity ) which is an achievement in it self . example input :- 168 ; output :- 3
example input :- 128 ; output :- 7